feat: reading position above/below (#15) + live options preview (#16)
- Add `readingPosition: 'above' | 'below'` setting; radio buttons in options page
- Extract shared rendering to `src/lib/panel.js` (STYLES, buildTokenGrid, buildPanel)
- Live preview card in options page renders pre-computed tokens in Shadow DOM,
  re-renders instantly on every input change

Closes #15, #16
