+++ src/content/renderer.js
@@ -1,167 +1,4 @@
-const STYLES = `
-  :host { display: block !important; }
-
-  /* ── Font size — driven by --panel-font-size custom property (#10) ── */
-  :host { --panel-font-size: 14px; }
-  :host([data-size="small"])  { --panel-font-size: 11px; }
-  :host([data-size="large"])  { --panel-font-size: 17px; }
-
-  /* ── Light mode (default) ── */
-  .panel {
-    margin: 8px 0 4px;
-    padding: 10px 14px;
-    border-left: 3px solid rgba(29, 155, 240, 0.55);
-    background: rgba(29, 155, 240, 0.05);
-    border-radius: 0 6px 6px 0;
-    font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, sans-serif;
-    font-size: var(--panel-font-size);
-    box-sizing: border-box;
-  }
-
-  /* ── Dark mode — applied when host has data-dark attribute ── */
-  :host([data-dark]) .panel {
-    background: rgba(29, 155, 240, 0.08);
-    border-left-color: rgba(29, 155, 240, 0.65);
-  }
-  :host([data-dark]) .token-surface { color: #e7e9ea; }
-  :host([data-dark]) .token-reading { color: rgb(29, 155, 240); }
-  :host([data-dark]) .token-romaji  { color: #71767b; }
-  :host([data-dark]) .token-plain   { color: #8b98a5; }
-  :host([data-dark]) .sep           { background: rgba(29, 155, 240, 0.2); }
-  :host([data-dark]) .value         { color: #e7e9ea; }
-  :host([data-dark]) .label         { color: rgb(29, 155, 240); }
-  :host([data-dark]) .loading       { color: #71767b; }
-  :host([data-dark]) .error         { color: #f4212e; }
-  :host([data-dark]) .note          { color: #71767b; }
-
-  /* ── Word-by-word token grid ── */
-
-  .token-grid {
-    display: flex;
-    flex-wrap: wrap;
-    align-items: flex-end;
-    gap: 2px 1px;
-    padding-bottom: 8px;
-  }
-
-  /* A word cell stacks: kanji / hiragana reading / romaji */
-  .token-cell {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    padding: 1px 4px;
-    min-width: 18px;
-  }
-
-  .token-surface {
-    font-size: 15px;
-    color: #0f1419;
-    line-height: 1.4;
-    text-align: center;
-  }
-
-  /* Hiragana reading — only shown when surface has kanji */
-  .token-reading {
-    font-size: 9.5px;
-    color: rgba(29, 155, 240, 0.9);
-    line-height: 1.3;
-    text-align: center;
-    font-weight: 500;
-  }
-
-  /* Romaji — shown for all Japanese tokens */
-  .token-romaji {
-    font-size: 9px;
-    color: #8899a6;
-    font-style: italic;
-    line-height: 1.3;
-    text-align: center;
-  }
-
-  /* Punctuation / Latin / numbers — inline, baseline-aligned */
-  .token-plain {
-    font-size: 15px;
-    color: #536471;
-    align-self: flex-end;
-    padding-bottom: 3px;
-    white-space: pre-wrap;
-  }
-
-  /* Forces a new line in the flex layout */
-  .token-break {
-    flex-basis: 100%;
-    height: 6px;
-  }
-
-  /* ── Translation rows ── */
-
-  .sep {
-    height: 1px;
-    background: rgba(29, 155, 240, 0.15);
-    margin: 2px 0 6px;
-  }
-
-  .row {
-    display: flex;
-    gap: 10px;
-    flex-wrap: wrap;
-    padding: 2px 0;
-    font-size: 13px;
-    line-height: 1.55;
-    color: #536471;
-  }
-
-  .label {
-    font-size: 10px;
-    font-weight: 700;
-    text-transform: uppercase;
-    letter-spacing: 0.6px;
-    color: rgba(29, 155, 240, 0.85);
-    min-width: 54px;
-    padding-top: 3px;
-    flex-shrink: 0;
-  }
-
-  .value {
-    flex: 1;
-    word-break: break-word;
-  }
-
-  /* ── Click-to-translate button (#6) ── */
-  .translate-btn {
-    margin-top: 4px;
-    padding: 4px 14px;
-    border: 1px solid rgba(29, 155, 240, 0.5);
-    border-radius: 9999px;
-    background: transparent;
-    color: rgb(29, 155, 240);
-    font-size: 12px;
-    font-weight: 600;
-    cursor: pointer;
-    transition: background 0.15s;
-  }
-  .translate-btn:hover { background: rgba(29, 155, 240, 0.1); }
-
-  /* ── States ── */
-
-  .loading {
-    color: #9aa5b1;
-    font-style: italic;
-    font-size: 12px;
-  }
-
-  .error {
-    font-size: 12px;
-    color: #e0245e;
-  }
-
-  .note {
-    font-size: 11px;
-    color: #9aa5b1;
-    font-style: italic;
-    margin-top: 4px;
-  }
-`;
+import { STYLES, buildTokenGrid, buildPanel } from '../lib/panel.js';
 
 /**
  * Detect X's dark mode by sampling the page background colour.
@@ -229,110 +66,22 @@ export function createAnnotationHost(tweetRoot, settings = {}) {
 /**
  * @param {HTMLElement} host
  * @param {Array}  tokens  — from getTokens()
- * @param {{ en, fr, truncated }} translations
+ * @param {{ lang1, lang2, truncated }} translations
+ * @param {object} settings
  */
-// Human-readable language names for the translation row labels
-const LANG_NAMES = {
-  en: 'English', fr: 'French',  es: 'Spanish',  de: 'German',
-  pt: 'Portuguese', it: 'Italian', nl: 'Dutch', pl: 'Polish',
-  ru: 'Russian', sv: 'Swedish', tr: 'Turkish', ar: 'Arabic',
-  zh: 'Chinese', ko: 'Korean',
-};
-
-export function updateAnnotation(host, tokens, { lang1, lang2, truncated }, settings = {}) {
-  const {
-    showReading = true,
-    showRomaji  = true,
-    showLang1   = true,
-    showLang2   = true,
-  } = settings;
+export function updateAnnotation(host, tokens, translations, settings = {}) {
   if (!host?.shadowRoot) return;
 
-  const panel = host.shadowRoot.querySelector('.panel');
-  panel.textContent = '';
+  const shadow = host.shadowRoot;
+  const style = shadow.querySelector('style');
 
-  // Token grid — word-by-word display
-  const grid = document.createElement('div');
-  grid.className = 'token-grid';
+  // Replace panel with freshly built one
+  const oldPanel = shadow.querySelector('.panel');
+  const newPanel = buildPanel(tokens, translations, settings);
+  shadow.replaceChild(newPanel, oldPanel);
 
-  for (const token of tokens) {
-    if (token.type === 'break') {
-      const br = document.createElement('div');
-      br.className = 'token-break';
-      grid.appendChild(br);
-      continue;
-    }
-
-    if (token.type === 'plain') {
-      const span = document.createElement('span');
-      span.className = 'token-plain';
-      span.textContent = token.surface;
-      grid.appendChild(span);
-      continue;
-    }
-
-    // type === 'word': kanji / hiragana reading / romaji
-    const cell = document.createElement('div');
-    cell.className = 'token-cell';
-
-    const surface = document.createElement('div');
-    surface.className = 'token-surface';
-    surface.textContent = token.surface;
-    cell.appendChild(surface);
-
-    if (token.reading && showReading) {
-      const reading = document.createElement('div');
-      reading.className = 'token-reading';
-      reading.textContent = token.reading;
-      cell.appendChild(reading);
-    }
-
-    if (token.romaji && showRomaji) {
-      const romaji = document.createElement('div');
-      romaji.className = 'token-romaji';
-      romaji.textContent = token.romaji;
-      cell.appendChild(romaji);
-    }
-
-    grid.appendChild(cell);
-  }
-
-  panel.appendChild(grid);
-
-  // Separator
-  const sep = document.createElement('div');
-  sep.className = 'sep';
-  panel.appendChild(sep);
-
-  // Translations — only render rows that are enabled (#3/#4)
-  const translationRows = [
-    { label: LANG_NAMES[settings.lang1] || settings.lang1 || 'English', value: lang1, show: showLang1 },
-    { label: LANG_NAMES[settings.lang2] || settings.lang2 || 'French',  value: lang2, show: showLang2 },
-  ].filter((r) => r.show && r.value && r.value !== '[translation unavailable]');
-
-  for (const { label, value } of translationRows) {
-    const row = document.createElement('div');
-    row.className = 'row';
-
-    const lbl = document.createElement('span');
-    lbl.className = 'label';
-    lbl.textContent = label;
-
-    const val = document.createElement('span');
-    val.className = 'value';
-    val.textContent = value;
-
-    row.appendChild(lbl);
-    row.appendChild(val);
-    panel.appendChild(row);
-  }
-
-  if (truncated) {
-    const note = document.createElement('div');
-    note.className = 'note';
-    note.textContent = 'Text was too long — translation may be partial.';
-    panel.appendChild(note);
-  }
+  // Re-attach style if it was removed
+  if (!shadow.querySelector('style')) shadow.prepend(style);
 }
 
 /**
@@ -341,49 +90,12 @@ export function updateAnnotation(host, tokens, { lang1, lang2, truncated }, sett
  */
 export function showTranslateButton(host, tokens, settings, onTranslate) {
   if (!host?.shadowRoot) return;
-  const panel = host.shadowRoot.querySelector('.panel');
-  panel.textContent = '';
+  const shadow = host.shadowRoot;
 
-  const { showReading = true, showRomaji = true } = settings;
-  const grid = document.createElement('div');
-  grid.className = 'token-grid';
-
-  for (const token of tokens) {
-    if (token.type === 'break') {
-      const br = document.createElement('div');
-      br.className = 'token-break';
-      grid.appendChild(br);
-      continue;
-    }
-    if (token.type === 'plain') {
-      const span = document.createElement('span');
-      span.className = 'token-plain';
-      span.textContent = token.surface;
-      grid.appendChild(span);
-      continue;
-    }
-    const cell = document.createElement('div');
-    cell.className = 'token-cell';
-    const surface = document.createElement('div');
-    surface.className = 'token-surface';
-    surface.textContent = token.surface;
-    cell.appendChild(surface);
-    if (token.reading && showReading) {
-      const reading = document.createElement('div');
-      reading.className = 'token-reading';
-      reading.textContent = token.reading;
-      cell.appendChild(reading);
-    }
-    if (token.romaji && showRomaji) {
-      const romaji = document.createElement('div');
-      romaji.className = 'token-romaji';
-      romaji.textContent = token.romaji;
-      cell.appendChild(romaji);
-    }
-    grid.appendChild(cell);
-  }
-
-  panel.appendChild(grid);
+  const oldPanel = shadow.querySelector('.panel');
+  const panel = document.createElement('div');
+  panel.className = 'panel';
+  panel.appendChild(buildTokenGrid(tokens, settings));
 
   const btn = document.createElement('button');
   btn.className = 'translate-btn';
@@ -397,6 +109,8 @@ export function showTranslateButton(host, tokens, settings, onTranslate) {
     onTranslate();
   }, { once: true });
   panel.appendChild(btn);
+
+  shadow.replaceChild(panel, oldPanel);
 }
 
 export function showAnnotationError(host, message = 'Translation failed.') {
